+++ libopendmi/src/entity/ipmi-device.c
@@ -7,7 +7,7 @@
 #include <limits.h>
 
 #include <opendmi/context.h>
-#include <opendmi/name.h>
+#include <opendmi/utils/name.h>
 #include <opendmi/utils.h>
 #include <opendmi/utils/decode.h>