+++ libopendmi/src/entity/system-boot.c
@@ -7,7 +7,7 @@
 #include <memory.h>
 
 #include <opendmi/context.h>
-#include <opendmi/name.h>
+#include <opendmi/utils/name.h>
 #include <opendmi/utils.h>
 
 #include <opendmi/entity/system-boot.h>